+++ src/me/bomb/cutscene/CameraManager_v1_15_R1.java
@@ -63,7 +63,7 @@ class CameraManager_v1_15_R1 extends CameraManager {
 	
 	static {
 		NonNullList<ItemStack> nnl = NonNullList.a();
-		for (int slot = 0; slot < 46; slot++) {
+		for (byte slot = 0; slot < 46; slot++) {
 			nnl.add(slot, new ItemStack(Item.getById(0)));
 		}
 		packetemptywindowitems = new PacketPlayOutWindowItems(0, nnl);
